requirements.txt
accelerate
controlnet-aux
diffusers
einops
einops
gguf
kornia
opencv-python
scikit-image
scipy
timm
torchsde
torchvision
transformers
xformers